requirements.txt
# tudo que eu preciar instalar para usar meu projeyo vai estar aqui
flask
mysql-connector-python



